move options icons
