--- redisinsight/ui/src/components/base/icons/options/ActiveActiveDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import ActiveActiveDark from 'uiSrc/assets/img/options/Active-ActiveDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const ActiveActiveDarkIcon = (props: IconProps) => (
-  <Icon icon={ActiveActiveDark} {...props} isSvg />
-)
