--- redisinsight/ui/src/components/base/icons/options/ActiveActiveLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import ActiveActiveLight from 'uiSrc/assets/img/options/Active-ActiveLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const ActiveActiveLightIcon = (props: IconProps) => (
-  <Icon icon={ActiveActiveLight} {...props} isSvg />
-)
