--- redisinsight/ui/src/components/base/icons/options/RedisOnFlashDarkIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisOnFlashDark from 'uiSrc/assets/img/options/RedisOnFlashDark.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisOnFlashDarkIcon = (props: IconProps) => (
-  <Icon icon={RedisOnFlashDark} {...props} isSvg />
-)
