--- redisinsight/ui/src/components/base/icons/options/RedisOnFlashLightIcon.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-import RedisOnFlashLight from 'uiSrc/assets/img/options/RedisOnFlashLight.svg?react'
-import { Icon, IconProps } from 'uiSrc/components/base/icons'
-
-export const RedisOnFlashLightIcon = (props: IconProps) => (
-  <Icon icon={RedisOnFlashLight} {...props} isSvg />
-)
