--- redisinsight/ui/src/components/base/icons/options/index.ts
@@ -1,4 +1,0 @@
-export { ActiveActiveDarkIcon } from './ActiveActiveDarkIcon'
-export { ActiveActiveLightIcon } from './ActiveActiveLightIcon'
-export { RedisOnFlashDarkIcon } from './RedisOnFlashDarkIcon'
-export { RedisOnFlashLightIcon } from './RedisOnFlashLightIcon'
